OAA-Backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
json5==0.9.25
pandas==2.2.3
psycopg2-binary==2.9.10
